+++ test/internal/Utils.t.sol
@@ -2,9 +2,10 @@
 pragma solidity ^0.8.20;
 
 import {Test} from "forge-std/Test.sol";
-import {strings} from "solidity-stringutils/src/strings.sol";
 
 import {Utils, ContractInfo} from "openzeppelin-foundry-upgrades/internal/Utils.sol";
+
+import {StringFinder} from "openzeppelin-foundry-upgrades/internal/StringFinder.sol";
 
 import {MyContractName} from "../contracts/MyContractFile.sol";
 
@@ -12,7 +13,7 @@ import {MyContractName} from "../contracts/MyContractFile.sol";
  * @dev Tests the Utils internal library.
  */
 contract UtilsTest is Test {
-    function testGetContractInfo_from_file() public view {
+    function testGetContractInfo_from_file() public {
         ContractInfo memory info = Utils.getContractInfo("Greeter.sol", "out");
 
         assertEq(info.shortName, "Greeter");
@@ -22,14 +23,14 @@ contract UtilsTest is Test {
         assertEq(info.sourceCodeHash, "0x9564e0245350d0eb5e42a8fed97d87518dbfbddf7668ed383f97a8558b2a9c39"); // source code hash of Greeter.sol
     }
 
-    function testGetContractInfo_from_fileAndName() public view {
+    function testGetContractInfo_from_fileAndName() public {
         ContractInfo memory info = Utils.getContractInfo("MyContractFile.sol:MyContractName", "out");
 
         assertEq(info.shortName, "MyContractName");
         assertEq(info.contractPath, "test/contracts/MyContractFile.sol");
     }
 
-    function testGetContractInfo_from_artifact() public view {
+    function testGetContractInfo_from_artifact() public {
         ContractInfo memory info = Utils.getContractInfo("out/MyContractFile.sol/MyContractName.json", "out");
 
         assertEq(info.shortName, "MyContractName");
@@ -48,7 +49,7 @@ contract UtilsTest is Test {
         }
     }
 
-    function testGetContractInfo_outDirTrailingSlash() public view {
+    function testGetContractInfo_outDirTrailingSlash() public {
         ContractInfo memory info = Utils.getContractInfo("Greeter.sol", "out/");
 
         assertEq(info.shortName, "Greeter");
@@ -62,19 +63,19 @@ contract UtilsTest is Test {
         } catch {}
     }
 
-    function testGetFullyQualifiedName_from_file() public view {
+    function testGetFullyQualifiedName_from_file() public {
         string memory fqName = Utils.getFullyQualifiedName("Greeter.sol", "out");
 
         assertEq(fqName, "test/contracts/Greeter.sol:Greeter");
     }
 
-    function testGetFullyQualifiedName_from_fileAndName() public view {
+    function testGetFullyQualifiedName_from_fileAndName() public {
         string memory fqName = Utils.getFullyQualifiedName("MyContractFile.sol:MyContractName", "out");
 
         assertEq(fqName, "test/contracts/MyContractFile.sol:MyContractName");
     }
 
-    function testGetFullyQualifiedName_from_artifact() public view {
+    function testGetFullyQualifiedName_from_artifact() public {
         string memory fqName = Utils.getFullyQualifiedName("out/MyContractFile.sol/MyContractName.json", "out");
 
         assertEq(fqName, "test/contracts/MyContractFile.sol:MyContractName");
@@ -103,7 +104,7 @@ contract UtilsTest is Test {
         assertEq(Utils.getOutDir(), "out");
     }
 
-    using strings for *;
+    using StringFinder for string;
 
     function testGetBuildInfoFile() public {
         ContractInfo memory contractInfo = Utils.getContractInfo("Greeter.sol", "out");
@@ -113,8 +114,8 @@ contract UtilsTest is Test {
             "out"
         );
 
-        assertTrue(buildInfoFile.toSlice().startsWith("out/build-info".toSlice()));
-        assertTrue(buildInfoFile.toSlice().endsWith(".json".toSlice()));
+        assertTrue(buildInfoFile.startsWith("out/build-info"));
+        assertTrue(buildInfoFile.endsWith(".json"));
     }
 
     function testToBashCommand() public pure {
@@ -133,11 +134,11 @@ contract UtilsTest is Test {
 }
 
 contract Invoker {
-    function getContractInfo(string memory contractName, string memory outDir) public view {
+    function getContractInfo(string memory contractName, string memory outDir) public {
         Utils.getContractInfo(contractName, outDir);
     }
 
-    function getFullyQualifiedName(string memory contractName, string memory outDir) public view {
+    function getFullyQualifiedName(string memory contractName, string memory outDir) public {
         Utils.getFullyQualifiedName(contractName, outDir);
     }
 }